feat: implement reading data groups

+++ shared/src/commonMain/kotlin/fr/outadoc/eidas/lds/ComData.kt
@@ -1,5 +1,5 @@
 package fr.outadoc.eidas.lds
 
 data class ComData(
-    val dataGroupNumbers: List<Int>,
+    val dataGroupNumbers: List<UByte>,
 )
